+++ app/src/main/java/woowacourse/payments/ui/newcard/components/NewCardContent.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard
+package woowacourse.payments.ui.newcard.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
